(coun): Last access used to report last login time. Counter-intuitive.  Report last connection time range instead. Unfortunately, this requires a change in output format.
(coun): Report on all countries even to mortals.  This used to be
done, and was most probably changed because it gave away too much
information.  Carefully limit the information given to mortals, as
follows.  Limit last access information to whether deities or allies
are logged on right now, no longer report connection time for deities
that are not logged on.  Remove column time for mortals, rename it to
BTU for deities.  Report status, but with less detail if HIDDEN is
enabled.