Revert "(prmptrd, uprmptrd): Should not be called while player->aborted.  Check."
This reverts commit fee8ac9d8f519649ad72129cae9ffb3bbd5d7ffb.

These *are* called while player->aborted.  Could be avoided, but: the
reason for not wanting to prompt then is to have each prompt consume a
line of input.  That's actually not feasible, because when we wait for
an argument (after prompting for it) when the update aborts commands,
we can't consume the argument we prompted for.