Move selector code from src/lib/global to src/lib/common
Future virtual selectors will need to access game state.  This depends
on common/file.c, which can't be used from global without creating a
cyclic dependency between libglobal.a and libcommon.a.

Move nsc.c to src/lib/common.  file.c depends on it, so move it as
well, renamed to filetable.c so it doesn't clash with the existing
file.c.