Remove bogus io_noblocking() from player_login()
The call switched the connection with the player to blocking I/O for
draining of output before closing the connection.  Looks scary,
because blocking on I/O blocks the complete server process, not just
the player thread.  But we don't do input, and we do output only with
IO_WAIT, which can't block.  So this has no effect.